services/test-harness/requirements.txt
httpx>=0.24,<1
opentelemetry-api>=1.20,<2
